feat: #61 프론트 페이지 구조 분리
